synthetic_data/generation_scripts/bedrock_requirements.txt
boto3>=1.42.0
python-dotenv>=1.0.0
